Create How to run this collection
First, you will need to fork the collection. Free code camp will give you step by step guide. https://www.freecodecamp.org/news/how-to-fork-a-github-repository/
As soon as you have forked the repo and imported to Postman, go to the three horizontal dots in front of the collection  and click on run collection.
That's it!